+++ test/browser.spec.js
@@ -16,11 +16,10 @@
 
 const utils = require('./utils.js');
 
-module.exports.addTests = function({testRunner, expect, defaultBrowserOptions, puppeteer}) {
+module.exports.addTests = function({testRunner, expect, puppeteer, headless}) {
   const {describe, xdescribe, fdescribe} = testRunner;
   const {it, fit, xit} = testRunner;
   const {beforeAll, beforeEach, afterAll, afterEach} = testRunner;
-  const headless = defaultBrowserOptions.headless;
 
   describe('Browser.version', function() {
     it('should return whether we are in headless', async({browser}) => {
@@ -46,41 +45,6 @@ module.exports.addTests = function({testRunner, expect, defaultBrowserOptions, p
       const remoteBrowser = await puppeteer.connect({browserWSEndpoint});
       expect(remoteBrowser.process()).toBe(null);
       await remoteBrowser.disconnect();
-    });
-  });
-
-  describe('Browser.Events.disconnected', function() {
-    it('should emitted when: browser gets closed, disconnected or underlying websocket gets closed', async() => {
-      const originalBrowser = await puppeteer.launch(defaultBrowserOptions);
-      const browserWSEndpoint = originalBrowser.wsEndpoint();
-      const remoteBrowser1 = await puppeteer.connect({browserWSEndpoint});
-      const remoteBrowser2 = await puppeteer.connect({browserWSEndpoint});
-
-      let disconnectedOriginal = 0;
-      let disconnectedRemote1 = 0;
-      let disconnectedRemote2 = 0;
-      originalBrowser.on('disconnected', () => ++disconnectedOriginal);
-      remoteBrowser1.on('disconnected', () => ++disconnectedRemote1);
-      remoteBrowser2.on('disconnected', () => ++disconnectedRemote2);
-
-      await Promise.all([
-        utils.waitEvent(remoteBrowser2, 'disconnected'),
-        remoteBrowser2.disconnect(),
-      ]);
-
-      expect(disconnectedOriginal).toBe(0);
-      expect(disconnectedRemote1).toBe(0);
-      expect(disconnectedRemote2).toBe(1);
-
-      await Promise.all([
-        utils.waitEvent(remoteBrowser1, 'disconnected'),
-        utils.waitEvent(originalBrowser, 'disconnected'),
-        originalBrowser.close(),
-      ]);
-
-      expect(disconnectedOriginal).toBe(1);
-      expect(disconnectedRemote1).toBe(1);
-      expect(disconnectedRemote2).toBe(1);
     });
   });
 };
