+++ test/cookies.spec.js
