+++ test/jshandle.spec.js
